Edited and ready to merge with dev
