+++ Web-Portal/models/blockly.py
@@ -1,5 +1,5 @@
 class Blockly:
-    def __init__(self, instructions=[]):
+    def __init__(self, instructions):
         self.__instructions = instructions
 
     def get_instructions(self):
