--- Web-Portal/templates/blockly.html
@@ -1,44 +1,0 @@
-{% extends "dashboard.html" %}
-
-{% block title %}blockly{% endblock %}
-
-{% block nav_blockly %}
-<li class="nav-item active">
-
-{% endblock %}
-
-{% block blockly %}
-<!DOCTYPE html>
-<html>
-
-<body>
-<p><b>You have <span id="capacity"></span> block(s) left.</b></p>
-  <p>
-    <button onclick="sendBlockValues()">Run Javascript</button>
-    <button onclick="showCode()">Show Javascript</button>
-  </p>
-
-  <div id="blocklyDiv" style="height: 480px; width: 600px;"></div>
-  <xml xmlns="https://developers.google.com/blockly/xml" id="toolbox" style="display: none">
-    <category name="Logic">
-    <block type="forward"></block>
-    <block type="Left"></block>
-    <block type="Right"></block>
-    <block type="controls_repeat_ext"></block>
-    <block type="math_number">
-      <field name="NUM">1</field>
-    </block>
-
-    </category>
-  </xml>
-
-  <input class="form-control" type="hidden" id="value" value="">
-<td style="width:50%">
-        <textarea id="code_output" readonly></textarea>
-</td>
-
-</body>
-</html>
-
-{% endblock %}
-
